+++ tasks/mpi/nikolaev_r_simple_iteration_method/src/ops_mpi.cpp
@@ -6,7 +6,7 @@
 
 bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodSequential::pre_processing() {
   internal_order_test();
-  int n = *reinterpret_cast<int*>(taskData->inputs[0]);
+  size_t n = *reinterpret_cast<size_t*>(taskData->inputs[0]);
   A_.assign(n * n, 0.0);
   b_.assign(n, 0.0);
   x_.assign(n, 1.0);
@@ -23,7 +23,7 @@ bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodSequential::va
     std::cerr << "Error: Incorrect amount of inputs and outputs." << std::endl;
     return false;
   }
-  int n = *reinterpret_cast<int*>(taskData->inputs[0]);
+  size_t n = *reinterpret_cast<int*>(taskData->inputs[0]);
   if (n < 1) {
     std::cerr << "Error: Incorrect matrix size." << std::endl;
     return false;
@@ -43,17 +43,17 @@ bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodSequential::va
 
 bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodSequential::run() {
   internal_order_test();
-  int n = b_.size();
+  size_t n = b_.size();
 
   std::vector<double> B(n * n, 0.0);
   std::vector<double> g(n, 0.0);
 
-  for (int i = 0; i < n; ++i) {
+  for (size_t i = 0; i < n; ++i) {
     if (A_[i * n + i] == 0) {
       std::cerr << "Error: Zero diagonal element detected in matrix A." << std::endl;
       return false;
     }
-    for (int j = 0; j < n; ++j) {
+    for (size_t j = 0; j < n; ++j) {
       if (i == j) {
         B[i * n + j] = 0;
       } else {
@@ -63,16 +63,16 @@ bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodSequential::ru
     g[i] = b_[i] / A_[i * n + i];
   }
   std::vector<double> x_old(n, 0.0);
-  for (int iter = 0; iter < max_iterations_; ++iter) {
+  for (size_t iter = 0; iter < max_iterations_; ++iter) {
     std::vector<double> x_new(n, 0.0);
-    for (int i = 0; i < n; ++i) {
+    for (size_t i = 0; i < n; ++i) {
       x_new[i] = g[i];
-      for (int j = 0; j < n; ++j) {
+      for (size_t j = 0; j < n; ++j) {
         x_new[i] += B[i * n + j] * x_old[j];
       }
     }
     double max_diff = 0.0;
-    for (int i = 0; i < n; ++i) {
+    for (size_t i = 0; i < n; ++i) {
       max_diff = std::max(max_diff, fabs(x_new[i] - x_old[i]));
     }
     if (max_diff < tolerance_) {
@@ -96,7 +96,7 @@ bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodSequential::po
 bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodParallel::pre_processing() {
   internal_order_test();
   if (world.rank() == 0) {
-    int n = *reinterpret_cast<int*>(taskData->inputs[0]);
+    size_t n = *reinterpret_cast<size_t*>(taskData->inputs[0]);
     A_.assign(n * n, 0.0);
     b_.assign(n, 0.0);
     x_.assign(n, 0.0);
@@ -115,7 +115,7 @@ bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodParallel::vali
       std::cerr << "Error: Incorrect amount of inputs and outputs." << std::endl;
       return false;
     }
-    int n = *reinterpret_cast<int*>(taskData->inputs[0]);
+    size_t n = *reinterpret_cast<size_t*>(taskData->inputs[0]);
     if (n < 1) {
       std::cerr << "Error: Incorrect matrix size." << std::endl;
       return false;
@@ -137,7 +137,7 @@ bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodParallel::vali
 bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodParallel::run() {
   internal_order_test();
 
-  int n = b_.size();
+  size_t n = b_.size();
   std::vector<double> local_A, local_b, local_x, x_prev(n, 0.0);
   std::vector<double> B(n * n, 0.0);
   std::vector<double> g(n, 0.0);
@@ -146,8 +146,8 @@ bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodParallel::run(
   std::vector<int> displs(world.size(), 0);
 
   if (world.rank() == 0) {
-    int base_size = n / world.size();
-    int remainder = n % world.size();
+    size_t base_size = n / world.size();
+    size_t remainder = n % world.size();
 
     for (int i = 0; i < world.size(); ++i) {
       sizes[i] = base_size + (i < remainder ? 1 : 0);
@@ -161,7 +161,7 @@ bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodParallel::run(
   boost::mpi::broadcast(world, displs, 0);
   boost::mpi::broadcast(world, n, 0);
 
-  int local_size = sizes[world.rank()];
+  size_t local_size = sizes[world.rank()];
   local_A.resize(local_size * n);
   local_b.resize(local_size);
   local_x.resize(local_size, 0.0);
@@ -169,13 +169,13 @@ bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodParallel::run(
   boost::mpi::scatterv(world, A_.data(), sizes, displs, local_A.data(), local_size * n, 0);
   boost::mpi::scatterv(world, b_.data(), sizes, displs, local_b.data(), local_size, 0);
 
-  for (int i = 0; i < n; ++i) {
+  for (size_t i = 0; i < n; ++i) {
     if (A_[i * n + i] == 0) {
       std::cerr << "Error: Zero diagonal element detected in matrix A." << std::endl;
       return false;
     }
 
-    for (int j = 0; j < n; ++j) {
+    for (size_t j = 0; j < n; ++j) {
       if (i == j) {
         B[i * n + j] = 0;
       } else {
@@ -186,16 +186,16 @@ bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodParallel::run(
     g[i] = b_[i] / A_[i * n + i];
   }
 
-  for (int iter = 0; iter < max_iterations_; ++iter) {
+  for (size_t iter = 0; iter < max_iterations_; ++iter) {
     boost::mpi::broadcast(world, x_prev, 0);
 
     std::vector<double> local_x_new(local_size, 0.0);
 
-    for (int i = 0; i < local_size; ++i) {
-      int global_index = displs[world.rank()] + i;
+    for (size_t i = 0; i < local_size; ++i) {
+      size_t global_index = displs[world.rank()] + i;
       local_x_new[i] = g[global_index];
 
-      for (int j = 0; j < n; ++j) {
+      for (size_t j = 0; j < n; ++j) {
         local_x_new[i] += B[global_index * n + j] * x_prev[j];
       }
     }
@@ -205,7 +205,7 @@ bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodParallel::run(
     if (world.rank() == 0) {
       double max_diff = 0.0;
 
-      for (int i = 0; i < n; ++i) {
+      for (size_t i = 0; i < n; ++i) {
         max_diff = std::max(max_diff, fabs(x_[i] - x_prev[i]));
       }
 
@@ -232,18 +232,18 @@ bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodParallel::post
 }
 
 bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodSequential::is_singular(const std::vector<double>& A,
-                                                                                          int n) {
+                                                                                          size_t n) {
   std::vector<std::vector<double>> mat(n, std::vector<double>(n));
-  for (int i = 0; i < n; ++i) {
-    for (int j = 0; j < n; ++j) {
+  for (size_t i = 0; i < n; ++i) {
+    for (size_t j = 0; j < n; ++j) {
       mat[i][j] = A[i * n + j];
     }
   }
   double det = 1;
-  for (int i = 0; i < n; ++i) {
+  for (size_t i = 0; i < n; ++i) {
     if (mat[i][i] == 0) {
       bool found = false;
-      for (int k = i + 1; k < n; ++k) {
+      for (size_t k = i + 1; k < n; ++k) {
         if (mat[k][i] != 0) {
           swap(mat[i], mat[k]);
           det *= -1;
@@ -255,9 +255,9 @@ bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodSequential::is
         return true;
       }
     }
-    for (int j = i + 1; j < n; ++j) {
+    for (size_t j = i + 1; j < n; ++j) {
       double ratio = mat[j][i] / mat[i][i];
-      for (int k = i; k < n; ++k) {
+      for (size_t k = i; k < n; ++k) {
         mat[j][k] -= ratio * mat[i][k];
       }
     }
@@ -267,12 +267,12 @@ bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodSequential::is
 }
 
 bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodSequential::is_diagonally_dominant(
-    const std::vector<double>& A, int n) {
-  for (int i = 0; i < n; ++i) {
+    const std::vector<double>& A, size_t n) {
+  for (size_t i = 0; i < n; ++i) {
     double diagonal_element = fabs(A[i * n + i]);
     double sum = 0.0;
 
-    for (int j = 0; j < n; ++j) {
+    for (size_t j = 0; j < n; ++j) {
       if (j != i) {
         sum += fabs(A[i * n + j]);
       }
@@ -286,18 +286,18 @@ bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodSequential::is
 }
 
 bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodParallel::is_singular(const std::vector<double>& A,
-                                                                                        int n) {
+                                                                                        size_t n) {
   std::vector<std::vector<double>> mat(n, std::vector<double>(n));
-  for (int i = 0; i < n; ++i) {
-    for (int j = 0; j < n; ++j) {
+  for (size_t i = 0; i < n; ++i) {
+    for (size_t j = 0; j < n; ++j) {
       mat[i][j] = A[i * n + j];
     }
   }
   double det = 1;
-  for (int i = 0; i < n; ++i) {
+  for (size_t i = 0; i < n; ++i) {
     if (mat[i][i] == 0) {
       bool found = false;
-      for (int k = i + 1; k < n; ++k) {
+      for (size_t k = i + 1; k < n; ++k) {
         if (mat[k][i] != 0) {
           swap(mat[i], mat[k]);
           det *= -1;
@@ -309,9 +309,9 @@ bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodParallel::is_s
         return true;
       }
     }
-    for (int j = i + 1; j < n; ++j) {
+    for (size_t j = i + 1; j < n; ++j) {
       double ratio = mat[j][i] / mat[i][i];
-      for (int k = i; k < n; ++k) {
+      for (size_t k = i; k < n; ++k) {
         mat[j][k] -= ratio * mat[i][k];
       }
     }
@@ -321,12 +321,12 @@ bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodParallel::is_s
 }
 
 bool nikolaev_r_simple_iteration_method_mpi::SimpleIterationMethodParallel::is_diagonally_dominant(
-    const std::vector<double>& A, int n) {
-  for (int i = 0; i < n; ++i) {
+    const std::vector<double>& A, size_t n) {
+  for (size_t i = 0; i < n; ++i) {
     double diagonal_element = fabs(A[i * n + i]);
     double sum = 0.0;
 
-    for (int j = 0; j < n; ++j) {
+    for (size_t j = 0; j < n; ++j) {
       if (j != i) {
         sum += fabs(A[i * n + j]);
       }
