+++ tasks/seq/nikolaev_r_simple_iteration_method/include/ops_seq.hpp
@@ -22,9 +22,9 @@ class SimpleIterationMethodSequential : public ppc::core::Task {
   std::vector<double> b_{};
   std::vector<double> x_{};
   double tolerance_ = 1e-3;
-  int max_iterations_ = 150;
+  size_t max_iterations_ = 150;
 
-  static bool is_singular(const std::vector<double>& A, int n);
-  static bool is_diagonally_dominant(const std::vector<double>& A, int n);
+  static bool is_singular(const std::vector<double>& A, size_t n);
+  static bool is_diagonally_dominant(const std::vector<double>& A, size_t n);
 };
 }  // namespace nikolaev_r_simple_iteration_method_seq
